requirements.txt
cffi==1.16.0
cryptography==42.0.5
mysql==0.0.3
mysqlclient==2.2.4
psycopg2==2.9.9
pycparser==2.21
python-dotenv==1.0.1
